+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# [2.1.2-beta.2]
+
+* Restore Keys from secret-key
+* Restore Keys from mnemonic and derivationPath.
+
 # [2.1.2-beta.1]
 
 * Operation Gas fee calculator optimisation